L104 from react samurai was made! Redo reducers with typescript.
