+++ package.json
@@ -7,6 +7,11 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.6.3",
+    "@types/jest": "^27.0.1",
+    "@types/node": "^16.7.10",
+    "@types/react": "^17.0.19",
+    "@types/react-dom": "^17.0.9",
+    "@types/redux-form": "^8.3.2",
     "axios": "^0.21.1",
     "classnames": "^2.3.1",
     "react": "^17.0.1",
@@ -18,6 +23,7 @@
     "redux-form": "^8.3.7",
     "redux-thunk": "^2.3.0",
     "reselect": "^4.0.0",
+    "typescript": "^4.4.2",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
