--- src/redux/sideBarReducer.js
@@ -1,8 +1,0 @@
-let initialState = {};
-
-// В reducer передаются action и state. state, относящийся к данной ветке
-const sideBarReducer = (state = initialState, action) => {
-    return state;
-};
-
-export default sideBarReducer;
